split server in multiple files
